small fixes

+++ src/main/java/at/hannibal2/skyhanni/config/features/garden/pests/PestTimerConfig.java
@@ -78,7 +78,7 @@ public class PestTimerConfig {
         desc = "Trigger the pest cooldown warning x seconds before cooldown ends."
     )
     @ConfigEditorSlider(minValue = 0, maxValue = 30, minStep = 1)
-    public float cooldownWarningTime = 0;
+    public int cooldownWarningTime = 0;
 
     @Expose
     @ConfigOption(
